Worked on Alloc and Dealloc functions.
These functions are almost finished. Having issue with the dealloc
function. Issues tend to pop up when I try to move through the memory as
if it is contiguous. Debug's log function needs an overhaul, since you
have to create the string and then pass it through the function. Also,
renamed the MemoryAllocators folder to FixedSizeAllocator, so that it
represented what it would actually be used for.