Made the Alloc() generic.
The Alloc() function now is generic and will save the data in the data
type passed into the template. It will also, return a pointer based on the
data type passed into the template. I did it this way, since it is the
easiest way to make it generic. I'm worried about the various versions of
the function created at compile time.